% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{TicagrelorVsClopidogrel_narrow}
\alias{TicagrelorVsClopidogrel_narrow}
\alias{TicagrelorVsClopidogrel_narrow-package}
\title{TicagrelorVsClopidogrel_narrow}
\description{
TicagrelorVsClopidogrel_narrow
}
